src/test/java/guru/springframework/spring6restmvc/bootstrap/BootstrapDataTest.java
package guru.springframework.spring6restmvc.bootstrap;

import guru.springframework.spring6restmvc.repositories.BeerRepository;
import guru.springframework.spring6restmvc.repositories.CustomerRepository;
import guru.springframework.spring6restmvc.services.BeerCsvService;
import guru.springframework.spring6restmvc.services.BeerCsvServiceImpl;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.core.io.Resource;

import static org.assertj.core.api.Assertions.assertThat;

@DataJpaTest
class BootstrapDataTest {

    private final BeerCsvService csvService = new BeerCsvServiceImpl();

    @Value("classpath:csvdata/beers.csv")
    private Resource beerResource;

    @Autowired
    private BeerRepository beerRepository;

    @Autowired
    private CustomerRepository customerRepository;

    private BootstrapData data;

    @BeforeEach
    void setUp() {
        data = new BootstrapData(beerRepository, customerRepository, csvService);
        data.setResource(beerResource);
    }

    @Test
    void testRun() {
        data.init();

        assertThat(beerRepository.count()).isEqualTo(2413);
        assertThat(customerRepository.count()).isEqualTo(3);
    }
}